Use presenter object for templates
The @page object is used in layouts to render various page parts such as
the title, meta tags, etc. A presenter object stands in for when this
isn't an instance of Page, so that it answers to the same methods
required by the template.

To present any generic model, just call present(@model) in your controller.

You can create custom presenters for your own models by inheriting from
BasePresenter.