Relabel and slightly refactor the image dialog size selector
The rationale here is that CSS values are easier to understand than
having to do a bunch of mental math with the 'modifier' values.

Also, this provides a little more UI help text display the label
of the thumbnail in the box, and having the raw pixel values in the
tooltip.