rename max_client_body_size configurable for Resources engine to max_file_size
max_client_body_size is not an accurate name for this option. We are not limiting the size of
your post body, we are limiting the maximum size of a file upload. max_client_body_size
should still be set on the webserver and be greater than the max_file_size.