Check to ensure render_with_templates? is defined when executing 404 page
Fixes #1795. When the error_404 method is called outside of the scope of the PagesController, as it can be, since extensions also are extended, too, the render_with_templates? call will throw a NoMethodError. This might also be rewritten as a redirect to the 404 page, which I think might be a better choice in the long run, but needs to be discussed.