Make multiple resource uploads pass on form parameters, which should be expected behaviour.
In the situation where you add a field to Resources, upon create, you're unable to save any of those additional attributes because we only pass on the file attribute at the moment when the user is uploading multiple files. Of course, the user is always uploading multiple files, because the default html5ish behaviour is to send an array anyway. This corrects the existing create method and allows the user to submit multiple attributes to be saved.