Update doc/guides/4 - Refinery Extensions/3 - Testing.textile
the above changes were needed by me to get the dummy generator running under 2.0.3. I'm not 100% sure on why the engine.rake/extension.rake switch was needed, but it was. 

And as of this commit
https://github.com/resolve/refinerycms/commit/9f215fc9b07c2c1e44064854781fb5e2c3e68921
it looks like we need `load_dummy_tasks` to get the, well, dummy tasks loaded.