Update menu_presenter.rb to get hide_children back and allow styling of ul
I updated the menu_presenter to have a condition with hide_children as I was using it before

I also added a list_tag_css which is usefull to style the ul (needed with twitter bootstrap nav)

I don't know if it is really interesting to integrate. It was just to let you know.