Vertical admin tabs
Having admin tabs at the top running horizontally means they have to fit into a finite space. If there are too many, they go into a carousel, which is very inconvenient for navigation, and makes the rightmost tabs invisible to users.

I've put the admin tabs vertically at the left of the admin area. This gives as much vertical space (with scrolling) as needed.

- positioned admin tabs vertically at left
- wrapped admin content in #admin_container div
- corresponding #login_container div for login page
- more rounded corner mixins
- box-sizing mixin
- admin_background mixing provides faux column gradient background for admin area (modern browsers and IE10 up)
- SASS variables for $menu_width and $standard_rounding
- sortable_menu.js setting changed to y axis
- removed jcarousel and images

Conflicts:
	core/app/assets/images/refinery/carousel-left.png
	core/app/assets/images/refinery/carousel-right.png
	core/app/assets/stylesheets/refinery/layout.css.scss