Removed spring from Gemfile, moved poltergeist and quiet_assets to groups.
Spring still works just fine for me if I put it in .gemfile which
is a file that we have for installing gems according to each developer's
preference.  If I completely uninstall Spring from my system then it
still works to invoke every command that asks for Spring due to
Spring rescuring from LoadError.

If I then place Spring back in .gemfile it will figure out that it
is now present and will load, speeding up subsequent invocations.