Add Matomo open web analytics (#3405)
This adds the possibility for a site to use the Matomo open web analytics (a.k.a. Piwik) for their web analytics instead of, or along with, Google Analytics. It can be enabled in the core config by setting the two required parameters `config.matomo_analytics_server` and `config.matomo_analytics_site_id`. 

It works analogue to the code for enabling Google Analytics and as such is pretty straightforward.